Move system test to components namespace
